pyproject.toml
[tool.poetry]
name = "python-tricks"
version = "0.1.0"
description = ""
authors = ["Gerrit Schrader <gerrit.schrader@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pykeepass = "^3.2.0"
attr = "^0.3.1"
python-box = "^4.2.3"
python-dotenv = "^0.14.0"
jinja2 = "^2.11.2"
pyyaml = "^5.3.1"
matplotlib = "^3.2.2"
jsonschema = "^3.2.0"
singleton_decorator = "^1.0.0"
pandas = "^1.0.5"
pydantic = "^1.5.1"
recordtype = "^1.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.3"
black = "^19.10b0"
mypy = "^0.782"

[tool.black]
line-length = 120
target-version = ['py38']
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

